Extract methods to prepare for testing
