+++ package.json
@@ -8,9 +8,9 @@
     "browserify": "browserify -r ./supermodal.js:SuperModal -s SuperModal > dist/supermodal.js",
     "html": "cp -v demo/index.html dist/index.html",
     "dist": "npm run scss && npm run browserify && npm run html",
-    "style": "eslint supermodal.js",
+    "style": "eslint supermodal.js lib/*.js",
     "watch:dist": "npm run dist && onchange 'supermodal.js' 'demo/*.html' 'scss/*.scss' -- npm run dist",
-    "watch:style": "npm run style & onchange 'supermodal.js' -- npm run style",
+    "watch:style": "npm run style & onchange 'supermodal.js' 'lib/**/*.js' -- npm run style",
     "serve": "npm run dist && browser-sync start --server dist --files 'dist/*'",
     "test": "npm run dist && karma start --single-run",
     "check": "npm run style"
@@ -35,6 +35,7 @@
     "eslint": "^0.13.0",
     "jasmine-core": "^2.2.0",
     "karma": "^0.12.31",
+    "karma-browserify": "^3.0.1",
     "karma-chrome-launcher": "^0.1.7",
     "karma-cli": "0.0.4",
     "karma-firefox-launcher": "^0.1.4",
